Updated Bandt-Pompe.R
Rewrote the ordinal pattern function so it is able to handle identical values. By assign each possible pattern with a weight = 1/possible_patterns. With a pattern with two identical values the two possible patterns will therefore both get weight 0.5. Made a script that can add all the weights together for each pattern. The function formationPatternMagnus outputs the sum of weights for each pattern. Taking the sum of all of the summed weights and dividing the ouput vector with this sum gives the opd. This is determinstic function